+++ scripts/create-local-demo.sh
@@ -6,6 +6,7 @@ cd "../cli"
 
 export COSMO_API_KEY=cosmo_669b576aaadc10ee1ae81d9193425705
 export COSMO_API_URL=http://localhost:3001
+export KC_API_URL=http://localhost:8080
 
 pnpm wgc federated-graph create production --label-matcher team=A,team=B --label-matcher env=production --routing-url http://localhost:3002/graphql
 
@@ -20,4 +21,4 @@ cd "../scripts"
 
 cd "../cli"
 
-pnpm wgc federated-graph create-token production --name mytoken
+pnpm wgc router token create mytoken --graph-name production